api/requirements.txt
Django==2.2.13
djangorestframework==3.11.1
django-rest-framework
django-cors-headers
flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
Pillow
PyJWT==1.6.4 
authentication



# gunicorn==19.9.0 # Our WSGI server !for Apache

